demo/requirements.txt
# Gemini Demo Backend Dependencies

# Flask web framework
flask>=2.3.0
flask-cors>=4.0.0

# Gemini SDK
google-generativeai>=0.3.0

# Environment variables
python-dotenv>=1.0.0

# Logging and utilities
python-json-logger>=2.0.7
